修复字典新增、编辑的问题

+++ coadmin-core/src/main/java/com/gitee/coadmin/modules/system/service/impl/DeptServiceImpl.java
@@ -216,7 +216,7 @@ public class DeptServiceImpl extends BaseServiceImpl<Dept> implements DeptServic
         return deptMapper.selectOne(wrapper);
     }
 
-    private int countOfChildren(Long pid) {
+    private long countOfChildren(Long pid) {
         QueryWrapper<Dept> wrapper = new QueryWrapper<>();
         wrapper.lambda().eq(Dept::getPid, pid);
         return deptMapper.selectCount(wrapper);
